requirements.txt
ctransformers
faiss-cpu
langchain
pypdf
python-box
sentence-transformers